Refactor: Extract Player into separate storyboard
- Renamed "Now Playing" to more adequate MiniPlayer
- Removed pan gesture from MiniPlayer to be added again later with a proper transition
- Tidied project structure with separe folders for views
