+++ BookPlayer/Views/BPMarqueeLabel.swift
